hacer nulos los campos del newrecipepost

+++ Business/InstructionProvider.cs
@@ -52,7 +52,7 @@ namespace NamNamAPI.Business
                     Cookinginstruction itemBD = new Cookinginstruction{
                         IdCookingInstruction = GenerateRandomID.GenerateID(),
                         Instruction = item.Instruction,
-                        Step = item.Step,
+                        Step = (int)item.Step,
                         RecipeIdRecipe = idRecipe
                     };
                     instructionsTemp.Add(itemBD);
